@@ -58,7 +58,7 @@
 		"rss": "^1.2.2",
 		"sass": "^1.75.0",
 		"shiki": "^1.3.0",
-		"svelte": "^4.2.15",
+		"svelte": "5.0.0-next.120",
 		"svelte-check": "^3.6.9",
 		"svelte-french-toast": "^1.2.0",
 		"svelte-markdown": "^0.4.1",
